Refactor: move WebGPU support warning into its own component.
